为了ios6的适配快吐了＝。＝

+++ MobileNJU/MobileNJU/一卡通/EcardVC.m
@@ -96,7 +96,7 @@
 {
     return 20;
 }
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 60;
 }
